chore: move miden wrapper functionality to a separate miden_wrapper file
